move cards directives
