+++ frontend/src/app/pages/game-view/current-investigator-panel/control-area/controlled-asset/asset-popover/asset-popover.component.ts
@@ -1,7 +1,7 @@
 import { ChangeDetectionStrategy, Component, input } from '@angular/core';
+import { CardBackgroundDirective } from '@shared/directives/cards/card-background.directive';
+import { CardOutlineDirective } from '@shared/directives/cards/card-outline.directive';
 import { PlayerCardBase } from 'shared/domain/entities/player-card.model';
-import { CardBackgroundDirective } from 'shared/ui/directives/card-background.directive';
-import { CardOutlineDirective } from 'shared/ui/directives/card-outline.directive';
 import { CardDetailsTextComponent } from '../../../../components/card-details-text/card-details-text.component';
 
 @Component({
