+++ frontend/src/app/pages/game-view/current-investigator-panel/investigator/investigator.component.ts
@@ -6,12 +6,12 @@ import {
   input,
 } from '@angular/core';
 import { TranslocoDirective } from '@jsverse/transloco';
+import { CardFactionBackgroundDirective } from '@shared/directives/cards/card-faction-background.directive';
+import { CardOutlineDirective } from '@shared/directives/cards/card-outline.directive';
 import { Investigator } from 'shared/domain/entities/investigator.model';
 import { CardInfoService } from 'shared/services/card-info.service';
 import { ImagesUrlService } from 'shared/services/images-url.service';
 import { VitalsBarComponent } from '../../../../shared/ui/components/vitals-bar/vitals-bar.component';
-import { CardFactionBackgroundDirective } from '../../../../shared/ui/directives/card-faction-background.directive';
-import { CardOutlineDirective } from '../../../../shared/ui/directives/card-outline.directive';
 import { InvestigatorAbilityComponent } from './investigator-ability/investigator-ability.component';
 import { InvestigatorSkillsComponent } from './investigator-skills/investigator-skills.component';
 import { InvestigatorSubtitleComponent } from './investigator-subtitle/investigator-subtitle.component';
