+++ frontend/src/app/pages/game-view/play-area/location/investigator-avatar/investigator-avatar.component.ts
@@ -9,8 +9,8 @@ import { Investigator } from '@domain/entities/investigator.model';
 import { AssetDetailIconComponent } from '@pages/game-view/current-investigator-panel/control-area/controlled-asset/asset-detail-icon/asset-detail-icon.component';
 import { ImagesUrlService } from '@services/images-url.service';
 import { VitalsBarComponent } from '@shared/components/vitals-bar/vitals-bar.component';
-import { CardFactionBackgroundDirective } from '@shared/directives/card-faction-background.directive';
-import { CardOutlineDirective } from '@shared/directives/card-outline.directive';
+import { CardFactionBackgroundDirective } from '@shared/directives/cards/card-faction-background.directive';
+import { CardOutlineDirective } from '@shared/directives/cards/card-outline.directive';
 
 @Component({
   selector: 'ah-investigator-avatar',
