+++ frontend/src/app/shared/ui/components/cards/card-parts/card-player-base/player-card.component.ts
@@ -6,11 +6,11 @@ import {
   input,
 } from '@angular/core';
 import { TranslocoPipe } from '@jsverse/transloco';
+import { CardOutlineDirective } from '@shared/directives/cards/card-outline.directive';
 import { cardHeights, cardWidths } from 'shared/domain/card.constants';
 import { DisplayOptions } from 'shared/domain/display.options';
 import { PlayerCardBase } from 'shared/domain/entities/player-card.model';
 import { ImagesUrlService } from 'shared/services/images-url.service';
-import { CardOutlineDirective } from 'shared/ui/directives/card-outline.directive';
 import { CardSkillsComponent } from '../card-skills/card-skills.component';
 
 @Component({
