+++ frontend/src/app/shared/ui/directives/cards/card-faction-background.directive.ts
@@ -1,6 +1,6 @@
 import { Directive, effect, ElementRef, inject, input } from '@angular/core';
-import { Faction } from '../../domain/entities/player-card.model';
-import { ImagesUrlService } from '../../services/images-url.service';
+import { Faction } from '@domain/entities/player-card.model';
+import { ImagesUrlService } from '@services/images-url.service';
 
 @Directive({
   selector: '[ahCardFactionBackground]',
